+++ backend/src/api-oauth/dto/oauth-token-response.dto.ts
@@ -2,6 +2,6 @@ export class OAuthTokenResponseDto {
     access_token: string;
     token_type: "bearer";
     expires_in: number;
-    refresh_token: string;
+    refresh_token?: string;
     scope: string;
 }
